add Card.js, CardList.js and cards.js and import them
