+++ js-structure/js-structure_quiz-app/index.js
@@ -1,31 +1,9 @@
 import Header from "./components/Header/Header.js";
 import Form from "./components/Form/Form.js";
 import Bookmark from "./components/Bookmark/Bookmark.js";
-
-// Store the cards in a global state
-const cards = [
-  {
-    question:
-      "In the Kingdom Hearts series who provides the english voice for Master Eraqus?",
-    answer: "Mark Hamill",
-    tags: ["kingdomheart", "mastereraqus", "english voices"],
-    isBookmarked: false,
-  },
-  {
-    question:
-      "Daniel Radcliffe became a global star in the film industry due to his performance in which film franchise?",
-    answer: "Harry Potter",
-    tags: ["hewhomustnotbenamed", "danielradcliffe", "abracadabra"],
-    isBookmarked: false,
-  },
-  {
-    question:
-      'In the game series "The Legend of Zelda" what is the name of the main character?',
-    answer: "Link",
-    tags: ["linktothepast", "linkbetweenworlds", "linkbetweenages"],
-    isBookmarked: false,
-  },
-];
+import Card from "./components/Card/Card.js";
+import CardList from "./components/CardList/CardList.js";
+import {cards} from "./utils/cards.js";
 
 
 const form = Form()
@@ -54,73 +32,6 @@ function handleFormSubmit(event) {
   event.target.elements.question.focus();
 }
 
-
-
-
-function Card(props) {
-  const card = document.createElement("article");
-  card.classList.add("card");
-  card.innerHTML = /* html */ `
-		<h2 class="card__question" data-js="question"></h2>
-		<button class="card__button-answer" type="button" data-js="answer-button">
-			Show answer
-		</button>
-		<p class="card__answer" data-js="answer"></p>
-		<ul class="card__tag-list" data-js="tags"></ul>
-		<div class="card__button-bookmark" data-js="bookmark-button-wrapper"></div>
-	`;
-
-  const question = card.querySelector('[data-js="question"]');
-  const answerButton = card.querySelector('[data-js="answer-button"]');
-  const answer = card.querySelector('[data-js="answer"]');
-  const tags = card.querySelector('[data-js="tags"]');
-  const bookmarkButtonWrapper = card.querySelector(
-    '[data-js="bookmark-button-wrapper"]'
-  );
-
-  question.textContent = props.question;
-  answer.textContent = props.answer;
-
-  props.tags.forEach((tag) => {
-    const tagElement = document.createElement("li");
-    tagElement.classList.add("card__tag");
-    tagElement.textContent = "#" + tag;
-    tags.append(tagElement);
-  });
-
-  function handleBookmarkButtonClick(event) {
-    cards.find((card) => card.question === props.question).isBookmarked =
-      !props.isBookmarked;
-    // To prevent the need to rerender, we can just toggle the class here
-    event.currentTarget.classList.toggle("bookmark--active");
-  }
-
-  const bookmarkButton = Bookmark({
-    active: props.isBookmarked,
-    onClick: handleBookmarkButtonClick,
-  });
-  bookmarkButtonWrapper.append(bookmarkButton);
-
-  function handleAnswerButtonClick() {
-    answer.classList.toggle("card__answer--active");
-  }
-
-  answerButton.addEventListener("click", handleAnswerButtonClick);
-
-  return card;
-}
-
-function CardList() {
-  const cardList = document.createElement("section");
-  cardList.classList.add("card-list");
-
-  cards.forEach((card) => {
-    const cardElement = Card(card);
-    cardList.append(cardElement);
-  });
-
-  return cardList;
-}
 
 function App() {
   const app = document.createElement("main");
